+++ src/module/automatedPolymorpherModels.ts
@@ -26,7 +26,7 @@ export class Creature {
 
 export enum PolymorpherFlags {
   // IS_LOCAL = 'isLocal',
-  STORE_ON_ACTOR = 'storeonactor',
+  // STORE_ON_ACTOR = 'storeonactor',
   POLYMORPHERS = 'polymorphers',
   RANDOM = 'random',
   ORDERED = 'ordered',
@@ -48,6 +48,7 @@ export class PolymorpherData {
   number: number;
   defaultsummontype: string;
   compendiumid: string;
+  explicitname:string;
 }
 
 export class PolymorpherCompendiumData {
@@ -73,6 +74,7 @@ export class TransformOptionsGeneric {
   removeAE = false;
   keepAEOnlyOriginNotEquipment = false;
   transformTokens = true;
+  explicitName = ''
 }
 
 /* pf2e */